Testnet Competition Code Changes (#1817)
* switch to have payable functions with autofix

* stuff builds, but lint is probably failing

* fix linting issues

* fix solhint plugin payable's resolution

* last lint fix pass

* fix failure in core-contracts

* fix election module build failure

* fix tests

* ERC functions should not be affected by our decisions

* enable basic features on perps market

createAccount is protected by feature flag

* fix compile errors

* add multicall module to spot market cannonfile

* revert ownable module to avoid proxy redeployment

* add amount to collateral price

* change name to size

* preset

* remove price deviation

* change update fee

* Trusted Multicall Forwarder (#1827)

* add trusted forwarder

* update context and change linter

* remove payables

* add new solhint

* successful build

* test fixes

* missing imports

* test fix

* test fix

* lint fix

* spot and perps markets do not need to include

trusted forwarder provision

* remove amount from external func

* use amount instead of 1

* Change getPositionCollateral to only return amount.
- The lets users get a position's collateral amount without having to do a price update due to ERC7412

* comments

* remove payable from multicall in multicallmodule

---------

+++ markets/perps-market/contracts/modules/GlobalPerpsMarketModule.sol
@@ -9,8 +9,8 @@ import {GlobalPerpsMarketConfiguration} from "../storage/GlobalPerpsMarketConfig
 import {GlobalPerpsMarket} from "../storage/GlobalPerpsMarket.sol";
 import {IGlobalPerpsMarketModule} from "../interfaces/IGlobalPerpsMarketModule.sol";
 import {OwnableStorage} from "@synthetixio/core-contracts/contracts/ownership/OwnableStorage.sol";
-import "@synthetixio/core-contracts/contracts/errors/AddressError.sol";
-import "@synthetixio/core-contracts/contracts/errors/ParameterError.sol";
+import {AddressError} from "@synthetixio/core-contracts/contracts/errors/AddressError.sol";
+import {ParameterError} from "@synthetixio/core-contracts/contracts/errors/ParameterError.sol";
 
 /**
  * @title Module for global Perps Market settings.
@@ -26,7 +26,7 @@ contract GlobalPerpsMarketModule is IGlobalPerpsMarketModule {
      */
     function setMaxCollateralAmount(
         uint128 synthMarketId,
-        uint collateralAmount
+        uint256 collateralAmount
     ) external override {
         OwnableStorage.onlyOwner();
         GlobalPerpsMarketConfiguration.Data storage store = GlobalPerpsMarketConfiguration.load();
@@ -38,7 +38,9 @@ contract GlobalPerpsMarketModule is IGlobalPerpsMarketModule {
     /**
      * @inheritdoc IGlobalPerpsMarketModule
      */
-    function getMaxCollateralAmount(uint128 synthMarketId) external view override returns (uint) {
+    function getMaxCollateralAmount(
+        uint128 synthMarketId
+    ) external view override returns (uint256) {
         return GlobalPerpsMarketConfiguration.load().maxCollateralAmounts[synthMarketId];
     }
